+++ lib/albiruni.dart
@@ -3,5 +3,8 @@ library albiruni;
 
 export 'src/albiruni_base.dart';
 export 'src/model/subject.dart';
+export 'src/util/exceptions.dart';
+export 'src/model/day_time.dart';
+export 'src/util/extension_methods.dart';
 
 // TODO: Export any libraries intended for clients of this package.