+++ lib/src/util/exceptions.dart
@@ -1,0 +1,9 @@
+class NoSubjectsException implements Exception {
+  String? message;
+  NoSubjectsException({this.message});
+}
+
+class EmptyBodyException implements Exception {
+  String? message;
+  EmptyBodyException({this.message});
+}